Filter posts by pubdate as well as drafts
This allows me to schedule a post if it's out of draft and the pub date
is before or equal to todays date.

Make sure I handle timezones properly by converting everything to UTC.
I set today's date hours:seconds to zero because that's what the hours:seconds
are for pubDate.

Finally, I chose to format the UTC date so I can do a simple string comparison